Add Cloud9 Stack and utilities.
Adds a separate CDK environment to generate the Cloud9 CFN Stack.

Provides a make target to upload the stack to a public S3 object to be
consumed by attendees. (If permissions are granted to the active user.)

Adds a make target to the main CDK directory to bootstrap CDK for students
so that they can start deploying the stacks once Cloud9 launches.