Add Cognito extension to Website stack.
Adds several CFN Outputs to use as hooks for configuring Cognito.

There is bootstrapping work to set up Cognito that is not yet done, this just tells CFN to spin up the resources.

NOTE: This is a partial migration to using more TOML configuration. I would like to do this everywhere, but it's an unrelated change to this changeset. So I am doing this only partially to reduce merge conflicts and keep unrelated changes down.

See: https://github.com/aws-samples/busy-engineers-document-bucket/issues/7